Reordered some files in order to refactor a bit
