+++ YUViewLib/src/decoder/decoderVVDec.h
@@ -34,10 +34,10 @@
 
 #include <QLibrary>
 
-#include "common/FileInfo.h"
-#include "decoderBase.h"
-#include "externalHeader/vvdec/vvdec.h"
-#include "video/videoHandlerYUV.h"
+#include <common/FileInfo.h>
+#include <decoder/decoderBase.h>
+#include <decoder/externalHeader/vvdec/vvdec.h>
+#include <video/yuv/videoHandlerYUV.h>
 
 namespace decoder
 {
@@ -112,9 +112,9 @@ private:
 
   void allocateNewDecoder();
 
-  vvdecDecoder *decoder{nullptr};
-  vvdecAccessUnit* accessUnit{nullptr};
-  vvdecFrame *currentFrame{nullptr};
+  vvdecDecoder *   decoder{nullptr};
+  vvdecAccessUnit *accessUnit{nullptr};
+  vvdecFrame *     currentFrame{nullptr};
 
   // Try to get the next picture from the decoder and save it in currentHMPic
   bool getNextFrameFromDecoder();
@@ -132,4 +132,4 @@ private:
   LibraryFunctionsVVDec lib{};
 };
 
-}
+} // namespace decoder
