+++ YUViewLib/src/video/rgb/PixelFormatRGBGuess.cpp
