+++ YUViewLib/src/video/rgb/PixelFormatRGBGuess.h
