+++ YUViewLib/src/video/rgb/videoHandlerRGBCustomFormatDialog.cpp
