+++ YUViewLib/src/video/videoHandlerDifference.cpp
@@ -36,7 +36,7 @@
 #include <algorithm>
 
 #include <common/Functions.h>
-#include <video/videoHandlerYUV.h>
+#include <video/yuv/videoHandlerYUV.h>
 
 namespace video
 {
