+++ YUViewLib/src/video/videoHandlerResample.h
@@ -34,8 +34,8 @@
 
 #include <common/FileInfo.h>
 
-#include "videoHandler.h"
-#include "videoHandlerYUV.h"
+#include <video/videoHandler.h>
+#include <video/yuv/videoHandlerYUV.h>
 
 #include <QPointer>
